move upload script to main.py
